Added Comments
Added the descriptions for the methods and classes I did

+++ src/main/java/Domain/GameObjects/Goblin.java
@@ -1,12 +1,14 @@
 package Domain.GameObjects;
 
 public class Goblin extends Enemy {
+    // goblin class is faster than knight class so it should have a greater speed multiplier.
     double speedMultiplier = 1.25;
     public Goblin(int xPos, int yPos, String enemyType, int healthPoints, int speed) {
         super(xPos, yPos, enemyType, healthPoints, speed);
         this.speed *= speedMultiplier;
     }
-
+    // this method is used for calculating the damage the goblin class takes based on the projectile
+    // and returns the amount of damage taken.
     @Override
     protected int CalcDamage(Projectile projectile) {
         switch (projectile.type) {
@@ -15,9 +17,9 @@ public class Goblin extends Enemy {
             case "Magic":
                 return (int)(projectile.damage * 0.75);
             case "Artillery":
-                return (int)(projectile.damage);
+                return (projectile.damage);
             default:
-                return (int)(projectile.damage);
+                return (projectile.damage);
         }
     }
 }
